Refactorized game_ctrl - Use STL containers - Use Pointers - Fixed signed/unsigned issues
git-svn-id: https://svn.code.sf.net/p/extremetuxracer/code/trunk@284 0420edf4-82e4-42fc-9478-35b55e6d67a3

+++ font.cpp
@@ -148,7 +148,7 @@ int CFont::LoadFont (const string& name, const char *path) {
 	fonts.back()->FaceSize (18);
 	fonts.back()->CharMap (ft_encoding_unicode);
 
-	fontindex = fontindex + "[" + name + "]" + Int_StrN (fonts.size()-1);
+	fontindex = fontindex + "[" + name + "]" + Int_StrN ((int)fonts.size()-1);
 	return (int)fonts.size()-1;
 }
